+++ src/app/http/controllers/HomeController.php
@@ -1,0 +1,15 @@
+<?php namespace Redooor\Redminportal\App\Http\Controllers;
+
+class HomeController extends Controller {
+    
+	/**
+	 * Show the application welcome screen to the user.
+	 *
+	 * @return Response
+	 */
+	public function home()
+	{
+		return view('redminportal::pages.home');
+	}
+
+}