+++ src/resources/views/pages/404.blade.php
@@ -1,9 +1,9 @@
 @extends('redminportal::layouts.master')
 
 @section('content')
-    <h1>Oops, 404!</h1>
     <div class='row'>
         <div class='col-md-12'>
+            <h1>Oops, 404!</h1>
             <h2>You seem to be lost!</h2>
             <p>
                 The page you're accessing is not available or does not exist at all.<br/>