+++ src/resources/views/pages/home.blade.php
@@ -27,6 +27,7 @@
                         <li>Allow multiple Category images (issue #79).</li>
                         <li>Add Post Model and Controller (issue #73).</li>
                         <li>Revamp Translation implementation for Category, Media, Module, Portfolio, Product and Promotion (issue #85).</li>
+                        <li>Add Page model and Controller (issue #74).</li>
                         <li>More coming... (developing)</li>
                     </ol>
                     <h4>Bug fixes:</h4>